feat: puzzle page fetch now
